backend/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn==1.7.2
joblib
python-dotenv
requests
reportlab
sqlalchemy
alembic
